requirements.txt
# Database and SQL dependencies
sqlalchemy>=2.0.0
pyodbc>=4.0.39
pandas>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0

# LangGraph for pipeline orchestration
langgraph>=0.2.0

# LangSmith for tracking and monitoring
langsmith>=0.1.0

# Streamlit and visualization
streamlit>=1.28.0
plotly>=5.17.0

# Optional: Additional utilities
numpy>=1.24.0
openpyxl>=3.1.0  # For Excel file support if needed 